(core) Upgrade to v1.0.0 of the paho mqtt client.
The artifacts are provided by a fork of the official v1.0.0 which has
been published in a manner which allows SBT to consume it.  Upstream,
there is a fix in place, but it's not available in a stable release
until v1.0.1.  Once v1.0.1 is released, then this project can revert
to the official channels.